👌 IMPROVE: Changed type of ID to Long in MentionedUsers

+++ src/main/kotlin/com/glinboy/test/largedata/entity/MentionedUsers.kt
@@ -6,7 +6,7 @@ import javax.persistence.Id
 @Entity
 data class MentionedUsers(
 
-    @Id val id: Int,
+    @Id val id: Long,
     val type: String,
     val username: String,
     val displayname: String,
